chore(backend): create accounting/tigerbeetle folder
