+++ packages/backend/src/accounting/errors.ts
@@ -1,20 +1,9 @@
-import {
-  CreateAccountError as CreateAccountErrorCode,
-  CreateTransferError as CreateTransferErrorCode
-} from 'tigerbeetle-node'
 import { AccountId } from './utils'
 
 export class CreateAccountError extends Error {
   constructor(public code: number) {
     super('CreateAccountError code=' + code)
     this.name = 'CreateAccountError'
-  }
-}
-
-export class CreateTransferError extends Error {
-  constructor(public code: CreateTransferErrorCode) {
-    super()
-    this.name = 'CreateTransferError'
   }
 }
 
@@ -40,28 +29,6 @@ export enum TransferError {
 // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/explicit-module-boundary-types
 export const isTransferError = (o: any): o is TransferError =>
   Object.values(TransferError).includes(o)
-
-export function areAllAccountExistsErrors(
-  errors: CreateAccountErrorCode[]
-): boolean {
-  return areAllOfTypeAccountErrors(errors, [
-    CreateAccountErrorCode.exists_with_different_debits_pending,
-    CreateAccountErrorCode.exists_with_different_debits_posted,
-    CreateAccountErrorCode.exists_with_different_credits_pending,
-    CreateAccountErrorCode.exists_with_different_credits_posted,
-    CreateAccountErrorCode.exists
-  ])
-}
-
-export function areAllOfTypeAccountErrors(
-  errorsOccurred: CreateAccountErrorCode[],
-  errToVerify: CreateAccountErrorCode[]
-): boolean {
-  for (const occurred of errorsOccurred) {
-    if (!errToVerify.includes(occurred)) return false
-  }
-  return true
-}
 
 export class BalanceTransferError extends Error {
   constructor(public error: TransferError) {
